fix: feat removendo os de:para que antes do contrato eram necessarios.

+++ src/users/users.service.ts
@@ -1,46 +1,45 @@
-import { Injectable } from '@nestjs/common';
-import { User } from '@prisma/client';
-import { PrismaUserRepository } from 'src/repository/prisma/userRepository';
+import { UserRepository } from 'src/repository/contracts/usersContracts';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { UserResponseDto } from './dto/getUserById.dto';
+import { User } from 'src/@types/entities/entityUser';
 
 @Injectable()
 export class UsersService {
-  constructor(private readonly userRepository: PrismaUserRepository) {}
-
+  constructor(private readonly userRepository: UserRepository) {}
   async createUser(
     data: Omit<User, 'id' | 'createdAt' | 'updatedAt'>,
   ): Promise<User | null> {
-    return await this.userRepository.create(data);
+    return this.userRepository.create(data);
   }
 
   async getUserById(id: number): Promise<UserResponseDto | null> {
     const user = await this.userRepository.getById(id);
     if (!user) {
-      return null;
+      throw new NotFoundException('Usuário Não Encontrado');
     }
-    return {
-      id: user.id,
-      name: user.name,
-      wage: user.wage,
-      enterprise: user.enterprise,
-      isSelected: user.isSelected,
-      createdAt: user.createdAt.toISOString(),
-      updatedAt: user.updatedAt.toISOString(),
-    };
+    return user;
   }
 
-  async getAllUsers(page: number = 1, limit: number = 10): Promise<User[]> {
+  async getAllUsers(
+    page: number = 1,
+    limit: number = 10,
+  ): Promise<UserResponseDto[]> {
     const validPage = page < 1 ? 1 : page;
     const validLimit = limit < 1 ? 10 : limit;
     const skip = (validPage - 1) * validLimit;
-    return this.userRepository.getAll(skip, Number(validLimit));
+    const users = await this.userRepository.getAll(skip, validLimit);
+    return users;
   }
 
   async updateUser(id: number, data: Partial<User>): Promise<User> {
-    return await this.userRepository.update(id, data);
+    return this.userRepository.update(id, data);
   }
 
   async deleteUser(id: number): Promise<void> {
+    const user = await this.userRepository.getById(id);
+    if (!user) {
+      throw new NotFoundException('Usuário Não Encontrado');
+    }
     await this.userRepository.delete(id);
   }
 }
